refactor: remove old `HomePage` and `Header`
