+++ src/App.tsx
@@ -1,6 +1,5 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Layout from "./components/Layout";
-import HomePage from "./pages/HomePage";
 import StockScreener from "./pages/StockScreener";
 import CompanySearch from "./pages/CompanySearch";
 import AboutPage from "./pages/AboutPage";
@@ -9,7 +8,7 @@ import LoginPage from "./pages/LoginPage";
 import PrivateDatabasePage from "./pages/PrivateDatabasePage";
 import SupabaseTestPage from "./pages/SupabaseTestPage";
 import { AuthProvider } from "./contexts/AuthContext";
-import HomePageV2 from "./pages/HomePageV2";
+import HomePage from "./pages/HomePage";
 import NotFoundPage from "./pages/NotFoundPage";
 
 function App() {
@@ -17,8 +16,8 @@ function App() {
 		<AuthProvider>
 			<Router>
 				<Routes>
+					<Route path="/" element={<HomePage />} />
 					<Route path="/" element={<Layout />}>
-						<Route index element={<HomePage />} />
 						<Route path="/screener" element={<StockScreener />} />
 						<Route path="/search" element={<CompanySearch />} />
 						<Route path="/contact" element={<ContactPage />} />
@@ -26,7 +25,6 @@ function App() {
 						<Route path="/private" element={<PrivateDatabasePage />} />
 						<Route path="/supabase" element={<SupabaseTestPage />} />
 					</Route>
-					<Route path="/v2" element={<HomePageV2 />} />
 					<Route path="/about" element={<AboutPage />} />
 					<Route path="*" element={<NotFoundPage />} />
 				</Routes>
