+++ src/components/Header.tsx
@@ -1,110 +1,48 @@
-import type React from "react";
-import { useState } from "react";
-import { Link, useLocation } from "react-router-dom";
+import { Link, NavLink } from "react-router-dom";
 
-const Header: React.FC = () => {
-	const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
-	const location = useLocation();
+const NAV_LINKS = [
+	{ label: "Home", to: "/" },
+	{ label: "Screener", to: "/screener" },
+	{ label: "API", to: "/api" },
+	{ label: "About", to: "/about" },
+] as const;
 
-	const navigation = [
-		{ name: "Home", href: "/" },
-		{ name: "Stock Screener", href: "/screener" },
-		{ name: "Company Search", href: "/search" },
-		{ name: "Supabase", href: "/supabase" },
-		{ name: "Private DB", href: "/private" },
-		{ name: "Login", href: "/login" },
-		{ name: "About", href: "/about" },
-		{ name: "Contact", href: "/contact" },
-	];
-
-	const isActive = (href: string) => {
-		return location.pathname === href;
-	};
-
+function Header() {
 	return (
-		<header className="glass-strong sticky top-0 z-50 mb-8">
-			<div className="container">
-				<nav className="flex items-center justify-between py-6">
-					{/* Logo */}
-					<Link to="/" className="flex items-center gap-2">
-						<div className="w-8 h-8 rounded-lg bg-primary flex items-center justify-center">
-							<span className="text-primary-foreground font-bold text-sm">
-								IC
-							</span>
-						</div>
-						<span className="text-xl font-bold text-foreground">
-							Iron Capital
-						</span>
+		<header className="border-b border-border">
+			<div className="max-w-6xl mx-auto px-4">
+				<nav
+					aria-label="Main navigation"
+					className="flex items-center justify-between h-16"
+				>
+					<Link to="/" aria-label="Iron Capital home">
+						<img
+							src="/icon.svg"
+							alt="Iron Capital"
+							className="w-9 h-9 rounded-full object-cover"
+						/>
 					</Link>
 
-					{/* Desktop Navigation */}
-					<div className="hidden md:flex items-center gap-8">
-						{navigation.map((item) => (
-							<Link
-								key={item.name}
-								to={item.href}
-								className={`px-4 py-2 rounded-lg transition-all duration-300 ${
-									isActive(item.href)
-										? "bg-primary text-white"
-										: "text-text-secondary hover:text-text-primary hover:bg-glass-bg"
-								}`}
-							>
-								{item.name}
-							</Link>
-						))}
-					</div>
-
-					{/* Mobile menu button */}
-					<button
-						type="button"
-						onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
-						className="md:hidden p-2 rounded-lg glass hover:glass-strong transition-all duration-300"
-						aria-label="Toggle menu"
-					>
-						<div className="w-6 h-6 flex flex-col justify-center items-center">
-							<span
-								className={`w-5 h-05 bg-text-primary transform transition-all duration-300 ${
-									isMobileMenuOpen ? "rotate-45 translate-y-1" : ""
-								}`}
-							/>
-							<span
-								className={`w-5 h-05 bg-text-primary transition-all duration-300 my-1 ${
-									isMobileMenuOpen ? "opacity-0" : ""
-								}`}
-							/>
-							<span
-								className={`w-5 h-05 bg-text-primary transform transition-all duration-300 ${
-									isMobileMenuOpen ? "-rotate-45 -translate-y-1" : ""
-								}`}
-							/>
-						</div>
-					</button>
-				</nav>
-
-				{/* Mobile Navigation */}
-				{isMobileMenuOpen && (
-					<div className="md:hidden py-4 border-t border-glass-border">
-						<div className="flex flex-col gap-2">
-							{navigation.map((item) => (
-								<Link
-									key={item.name}
-									to={item.href}
-									onClick={() => setIsMobileMenuOpen(false)}
-									className={`px-4 py-3 rounded-lg transition-all duration-300 ${
-										isActive(item.href)
-											? "bg-primary text-white"
-											: "text-text-secondary hover:text-text-primary hover:bg-glass-bg"
-									}`}
+					<ul className="flex items-center gap-6 list-none m-0 p-0">
+						{NAV_LINKS.map(({ label, to }) => (
+							<li key={to}>
+								<NavLink
+									to={to}
+									className={({ isActive }) =>
+										isActive
+											? "text-sm text-foreground transition-colors"
+											: "text-sm text-muted-foreground hover:text-foreground transition-colors"
+									}
 								>
-									{item.name}
-								</Link>
-							))}
-						</div>
-					</div>
-				)}
+									{label}
+								</NavLink>
+							</li>
+						))}
+					</ul>
+				</nav>
 			</div>
 		</header>
 	);
-};
+}
 
 export default Header;
